# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "summary": "Update current user profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/device-tokens": {
      "post": {
        "operationId": "UsersController_registerDeviceToken",
        "summary": "Register or refresh a push device token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device token registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_removeDeviceToken",
        "summary": "Remove a push device token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveDeviceTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device token removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/avatar": {
      "post": {
        "operationId": "UsersController_uploadAvatar",
        "summary": "Upload current user avatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Avatar uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/products": {
      "get": {
        "operationId": "ProductsController_list",
        "summary": "List public products with filters",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search nameAr / nameEn / SKU (ILIKE)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isFeatured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isBestSeller",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Admin filter: active products only",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingImages",
            "required": false,
            "in": "query",
            "description": "Admin filter: products with no images",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "lowStock",
            "required": false,
            "in": "query",
            "description": "Admin filter: low stock products",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Admin sort",
            "schema": {
              "enum": [
                "newest",
                "name",
                "price",
                "stock"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/v1/products/{id}/reviews": {
      "get": {
        "operationId": "ReviewsController_listPublic",
        "summary": "List approved visible product reviews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/products/{productId}/reviews": {
      "post": {
        "operationId": "ReviewsController_create",
        "summary": "Submit a product review for a completed order item",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Tharaa Market API",
    "description": "Official **Version 1** REST API contract for Tharaa Market (Al Khafji).\n\n### Stability\nThis API is **frozen**. Breaking changes to URLs, request/response bodies, enums, pagination, auth, uploads, or error envelopes are not allowed in v1.\nFuture incompatible changes must ship as `/api/v2`.\n\n### Response envelope\nSuccess: `{ success: true, data, meta? }`\nError: `{ success: false, statusCode, message, errorCode?, details?, timestamp, path }`\n\n### Authentication\nUse `Authorization: Bearer <accessToken>`. Obtain tokens via `/auth/login`, `/auth/register`, `/auth/google`, or `/auth/refresh`.\nRefresh tokens rotate on every use.\n\n### Uploads\nMultipart form field name: `file`. Allowed MIME: jpeg, png, webp. Served under `/static/`.\n\n### Admin realtime (Socket.IO)\nNamespace `/admin` for staff dashboards. Authenticate with the JWT access token (`auth.token` or `Authorization: Bearer`).\nEvents: `admin:order_created`, `admin:order_updated`. Roles: ADMIN, MANAGER, EMPLOYEE.\nFull contract: `docs/ADMIN_INTEGRATION.md` (Realtime section). REST remains the source of truth.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access JWT from auth endpoints"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Always false for errors"
          },
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable error message"
          },
          "errorCode": {
            "type": "string",
            "example": "Bad Request",
            "description": "Optional machine-readable error code"
          },
          "details": {
            "type": "object",
            "description": "Optional extra details (e.g. validation messages)"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-14T00:00:00.000Z",
            "description": "ISO-8601 timestamp when the error was produced"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products",
            "description": "Request path that produced the error"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ApiSuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "Always true for success"
          },
          "data": {
            "type": "object",
            "description": "Payload returned by the endpoint"
          },
          "meta": {
            "type": "object",
            "description": "Optional pagination or extra metadata",
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Ahmed Ali"
          },
          "phone": {
            "type": "string",
            "example": "+966500000000"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "locale": {
            "type": "string",
            "enum": [
              "ar",
              "en"
            ],
            "example": "ar"
          }
        }
      },
      "DeviceTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "FCM / push device token"
          },
          "platform": {
            "type": "string",
            "example": "android"
          },
          "locale": {
            "type": "string",
            "enum": [
              "ar",
              "en"
            ]
          }
        },
        "required": [
          "token"
        ]
      },
      "RemoveDeviceTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "orderItemId",
          "rating"
        ]
      },
...
    }
  }
}